refactor: Renamed descriptionIndex.ts and descriptionScript.js to MainDescription.ts and MainDescriptionScript.js for clarity and maintainability.
