+++ src/content/description/MainDescription.ts
@@ -23,7 +23,7 @@ export async function fetchOriginalDescription(): Promise<string | null> {
         };
         window.addEventListener('ynt-description-data', handleDescription as EventListener);
         const script = document.createElement('script');
-        script.src = browser.runtime.getURL('dist/content/scripts/descriptionScript.js');
+        script.src = browser.runtime.getURL('dist/content/scripts/MainDescriptionScript.js');
         document.documentElement.appendChild(script);
     });
 }
